Move CreateNewToken, UpdateOrCreateToken and encodeTokenSecretData to client-go for a generic abstraction and client for a Bootstrap Token
