+++ cmd/kubeadm/app/cmd/BUILD
@@ -71,6 +71,7 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/version:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/util/flag:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
+        "//vendor/k8s.io/client-go/tools/token:go_default_library",
         "//vendor/k8s.io/client-go/util/cert:go_default_library",
     ],
 )
