+++ cmd/kubeadm/app/phases/bootstraptoken/node/BUILD
@@ -15,20 +15,13 @@ go_test(
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "tlsbootstrap.go",
-        "token.go",
-    ],
+    srcs = ["tlsbootstrap.go"],
     deps = [
         "//cmd/kubeadm/app/constants:go_default_library",
         "//cmd/kubeadm/app/util/apiclient:go_default_library",
-        "//cmd/kubeadm/app/util/token:go_default_library",
         "//pkg/apis/rbac/v1beta1:go_default_library",
-        "//pkg/bootstrap/api:go_default_library",
         "//pkg/util/version:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/rbac/v1beta1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
     ],
