+++ staging/BUILD
@@ -187,6 +187,7 @@ filegroup(
         "//staging/src/k8s.io/client-go/tools/record:all-srcs",
         "//staging/src/k8s.io/client-go/tools/reference:all-srcs",
         "//staging/src/k8s.io/client-go/tools/remotecommand:all-srcs",
+        "//staging/src/k8s.io/client-go/tools/token:all-srcs",
         "//staging/src/k8s.io/client-go/transport:all-srcs",
         "//staging/src/k8s.io/client-go/util/cert:all-srcs",
         "//staging/src/k8s.io/client-go/util/exec:all-srcs",
